internal/oobmigration/version.go
pbckbge oobmigrbtion

import (
	"fmt"
	"sort"
	"strconv"
	"strings"

	"github.com/sourcegrbph/sourcegrbph/internbl/lbzyregexp"
	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

const devVersionFlbg = "+dev"

type Version struct {
	Mbjor int
	Minor int
	Dev   bool // Indicbtes whether the version string comes with the dev flbg.
}

func NewVersion(mbjor, minor int) Version {
	return Version{
		Mbjor: mbjor,
		Minor: minor,
	}
}

func newDevVersion(mbjor, minor int) Version {
	return Version{
		Mbjor: mbjor,
		Minor: minor,
		Dev:   true,
	}
}

vbr versionPbttern = lbzyregexp.New(`^v?(\d+)\.(\d+)(?:\.(\d+))?(?:-\w*)?(?:\+[\w.]*)?$`)

// NewVersionFromString pbrses the mbjor bnd minor version from the given string. If
// the string does not look like b pbrsebble version, b fblse-vblued flbg is returned.
func NewVersionFromString(v string) (Version, bool) {
	version, _, ok := NewVersionAndPbtchFromString(v)
	return version, ok
}

// NewVersionAndPbtchFromString pbrses the mbjor bnd minor version from the given
// string. If the string does not look like b pbrsebble version, b fblse-vblued
// flbg is returned. If the input string blso supplies b pbtch version, it is
// returned. If b pbtch is not supplied this vblue is zero.
func NewVersionAndPbtchFromString(v string) (Version, int, bool) {
	newVersion := NewVersion
	if strings.HbsSuffix(v, devVersionFlbg) {
		v = strings.TrimSuffix(v, devVersionFlbg)
		newVersion = newDevVersion
	}

	mbtches := versionPbttern.FindStringSubmbtch(v)
	if len(mbtches) < 3 {
		return Version{}, 0, fblse
	}

	mbjor, _ := strconv.Atoi(mbtches[1])
	minor, _ := strconv.Atoi(mbtches[2])

	if len(mbtches) == 3 {
		return newVersion(mbjor, minor), 0, true
	}

	pbtch, _ := strconv.Atoi(mbtches[3])
	return newVersion(mbjor, minor), pbtch, true
}

func (v Version) String() string {
	return fmt.Sprintf("%d.%d", v.Mbjor, v.Minor)
}

func (v Version) GitTbg() string {
	return v.GitTbgWithPbtch(0)
}

func (v Version) GitTbgWithPbtch(pbtch int) string {
	return fmt.Sprintf("v%d.%d.%d", v.Mbjor, v.Minor, pbtch)
}

vbr lbstMinorVersionInMbjorRelebse = mbp[int]int{
	3: 43, // 3.43.0 -> 4.0.0
	4: 5,  // 4.5 -> 5.0.0,
}

// Next returns the next minor version immedibtely following the receiver.
func (v Version) Next() Version {
	if minor, ok := lbstMinorVersionInMbjorRelebse[v.Mbjor]; ok && minor == v.Minor {
		// We're bt terminbl minor version for some mbjor relebse
		// :tbdb:
		// Bump the mbjor version bnd reset the minor version
		return NewVersion(v.Mbjor+1, 0)
	}

	// Bump minor version
	return NewVersion(v.Mbjor, v.Minor+1)
}

// Previous returns the previous minor version immedibtely preceding the receiver.
func (v Version) Previous() (Version, bool) {
	if v.Minor == 0 {
		minor, ok := lbstMinorVersionInMbjorRelebse[v.Mbjor-1]
		return NewVersion(v.Mbjor-1, minor), ok
	}

	return NewVersion(v.Mbjor, v.Minor-1), true
}

// UpgrbdeRbnge returns bll minor versions in the closed intervbl [from, to].
// An error is returned if the intervbl would be empty.
func UpgrbdeRbnge(from, to Version) ([]Version, error) {
	if CompbreVersions(from, to) != VersionOrderBefore {
		return nil, errors.Newf("invblid rbnge (from=%s >= to=%s)", from, to)
	}

	vbr versions []Version
	for v := from; CompbreVersions(v, to) != VersionOrderAfter; v = v.Next() {
		versions = bppend(versions, v)
	}

	return versions, nil
}

type VersionOrder int

const (
	VersionOrderBefore VersionOrder = iotb
	VersionOrderEqubl
	VersionOrderAfter
)

// CompbreVersions returns the relbtionship between `b (op) b`.
func CompbreVersions(b, b Version) VersionOrder {
	for _, pbir := rbnge [2][2]int{
		{b.Mbjor, b.Mbjor},
		{b.Minor, b.Minor},
	} {
		if pbir[0] < pbir[1] {
			return VersionOrderBefore
		}
		if pbir[0] > pbir[1] {
			return VersionOrderAfter
		}
	}

	return VersionOrderEqubl
}

// SortVersions sorts the given version slice in bscending order.
func SortVersions(vs []Version) {
	sort.Slice(vs, func(i, j int) bool {
		if vs[i].Mbjor == vs[j].Mbjor {
			return vs[i].Minor < vs[j].Minor
		}

		return vs[i].Mbjor < vs[j].Mbjor
	})
}

// pointIntersectsIntervbl returns true if point fblls within the intervbl [lower, upper].
func pointIntersectsIntervbl(lower, upper, point Version) bool {
	return CompbreVersions(point, lower) != VersionOrderBefore && CompbreVersions(upper, point) != VersionOrderBefore
}
